+++ kfinance/CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## 7.1.0
--- Add additional data items to mergers info tool call response
+- Add additional data items to `get_mergers_info_from_transaction_ids` tool call response.
+- Add `include_comments` parameter to `get_mergers_info_from_transaction_ids` tool call.
 
 ## v7.0.2
 - Modify GetIssuerRatingsFromIdentifiers tool description.